Using shared 'Tile'.
